Make sure plugins used by the tests are built and copied to the lightdb_test target.
test/CMakeLists.txt:
Iterate over all of the subdirectories under "plugins/" and add
each as a target that lightdb_test depends on so that they are built
when the lightdb_test target is built. Pass INSTALL_DIR, which is set
to the test/plugins directory, to each plugin so that it can copy
itself directly to the correct location in the test directory. Set
BUILD_ALWAYS to true so that if changes are made to any of the plugin's
source files, they will be rebuilt.

plugins/blur/CMakeLists.txt:
Copy the shared library directly to the correct location in the test
directory.

plugins/yolo/CMakeLists.txt:
Ditto. Also add an include statement that was necessary for yolo to
build. The includes should probably eventually reference the current
project's source, but that can happen later.