Make Discretize actually scale the output video.
It was using the same dimensions as the input frame, but it should
be using the discretized dimensions so that the output video is correctly
scaled.

To avoid outputting an INFO log for every downsampled frame, I moved the
log statement behind and `if` statement so it only prints if the dimensions change.

It looks like the change to use the input frame dimensions happened in
0c23d0d.